L5: fix updated_at problem
Eloquent Update requires updated_at to either be in the fillable array or to have a guarded array
in the Modem. So we globally use a guarded field from now ! Please do not use fillable array
anymore, except adding the updated_at field. Otherwise all update-Observers will not be called