IpPool lifecycle test running (with caveats!).
Problem is a complex validation of used IP pools which is hard to
implement in the seeder – so I disabled testing of create using fake
data.
On the other hand it is possible to empty create (without fake data) as
the controller autofills the create form. Because of this update and
delete tests are running, too.